requirements.txt
streamlit
pandas
numpy
biopython
scikit-learn
